crm-web2/pyproject.toml
[tool.poetry]
name = "crm-web"
version = "0.1.0"
description = ""
authors = ["shigeokatakeshi <takeshi0713@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.1.2"
peewee = "^3.15.0"
python-dotenv = "^0.20.0"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
